refactor: encapsulate all the movie details logic into useMovie hook
